2. Recognising Faces/Source/requirements.txt
cv2-python
imutils
pandas
numpy
argparse
boto3
PIL